Move cluster methods to its own module
